requirements.txt
python-dotenv
instagrapi
jdatetime
pillow
rich